+++ src/AutorestSwift/Models/Complex/ConstantSchema.swift
@@ -28,6 +28,36 @@ import Foundation
 
 /// a schema that represents a constant value
 class ConstantSchema: Schema {
+    /// a container for the actual constant value
+    class ConstantValue: Codable, OptionalLanguageShortcut {
+        /// per-language information for this value
+        public var language: Languages?
+
+        /// the actual constant value to use
+        let value: String
+
+        /// Additional metadata extensions dictionary
+        let extensions: [String: AnyCodable]?
+
+        enum CodingKeys: String, CodingKey {
+            case language, value, extensions
+        }
+
+        public required init(from decoder: Decoder) throws {
+            let container = try decoder.container(keyedBy: CodingKeys.self)
+            language = try? container.decode(Languages?.self, forKey: .language)
+            value = try container.decode(String.self, forKey: .value)
+            extensions = try? container.decode([String: AnyCodable].self, forKey: .extensions)
+        }
+
+        public func encode(to encoder: Encoder) throws {
+            var container = encoder.container(keyedBy: CodingKeys.self)
+            if language != nil { try? container.encode(language, forKey: .language) }
+            try container.encode(value, forKey: .value)
+            if extensions != nil { try container.encode(extensions, forKey: .extensions) }
+        }
+    }
+
     /// the schema type of the constant value (ie, StringSchema, NumberSchema, etc)
     let valueType: Schema
