--- src/AutorestSwift/Models/Simple/ConstantValue.swift
@@ -1,57 +1,0 @@
-// --------------------------------------------------------------------------
-//
-// Copyright (c) Microsoft Corporation. All rights reserved.
-//
-// The MIT License (MIT)
-//
-// Permission is hereby granted, free of charge, to any person obtaining a copy
-// of this software and associated documentation files (the ""Software""), to
-// deal in the Software without restriction, including without limitation the
-// rights to use, copy, modify, merge, publish, distribute, sublicense, and/or
-// sell copies of the Software, and to permit persons to whom the Software is
-// furnished to do so, subject to the following conditions:
-//
-// The above copyright notice and this permission notice shall be included in
-// all copies or substantial portions of the Software.
-//
-// THE SOFTWARE IS PROVIDED *AS IS*, WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING
-// FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS
-// IN THE SOFTWARE.
-//
-// --------------------------------------------------------------------------
-
-import Foundation
-
-/// a container for the actual constant value
-class ConstantValue: Codable, OptionalLanguageShortcut {
-    /// per-language information for this value
-    public var language: Languages?
-
-    /// the actual constant value to use
-    let value: String
-
-    /// Additional metadata extensions dictionary
-    let extensions: [String: AnyCodable]?
-
-    enum CodingKeys: String, CodingKey {
-        case language, value, extensions
-    }
-
-    public required init(from decoder: Decoder) throws {
-        let container = try decoder.container(keyedBy: CodingKeys.self)
-        language = try? container.decode(Languages?.self, forKey: .language)
-        value = try container.decode(String.self, forKey: .value)
-        extensions = try? container.decode([String: AnyCodable].self, forKey: .extensions)
-    }
-
-    public func encode(to encoder: Encoder) throws {
-        var container = encoder.container(keyedBy: CodingKeys.self)
-        if language != nil { try? container.encode(language, forKey: .language) }
-        try container.encode(value, forKey: .value)
-        if extensions != nil { try container.encode(extensions, forKey: .extensions) }
-    }
-}
