+++ templates/OperationOptionsDecodingSnippet.stencil
@@ -3,6 +3,6 @@ let {{ param.key }}String = DateFormatter().string(from: {{ param.key }})
 {% elif param.strategy == "byteArrayFromParam" %}
 let {{ param.key }}String = String(bytes: {{ param.key }}, encoding: .utf8)
 {% else %}
-// TODO: Couldn't find template for {{ param.strategy }}
+{# No need to decode the parameter to String. Leave this blank #}
 {% endif %}
