+++ templates/ServiceClientFile.stencil
@@ -75,8 +75,8 @@ public final class {{ model.name }}: {{ model.protocols }} {
     public var {{ param.name }}: {{ param.type }}{{ param.defaultValue }}
 {% endfor %}
 
-{% for property in model.namedOperationGroups.values %}
-    private(set) lazy var {{ property.name.lowercasedFirst }}: {{ property.name }} = {{ property.name }}(client: self)
+{% for property in model.namedOperationGroupShortcuts %}
+    private(set) lazy var {{ property.value }}: {{ property.key }} = {{ property.key }}(client: self)
 {%endfor %}
 
     // MARK: Public Client Methods
